+++ CHANGELOG.md
@@ -13,6 +13,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Removed
 - None.
 ### Fixed
-- None.
+- Updated project to use URLs instead of String paths. https://github.com/JamitLabs/ProjLint/pull/33
 ### Security
 - None.